requirements.txt
# PyTorch with CUDA 11.8
torch==2.2.2+cu118
torchvision==0.17.2+cu118
torchaudio==2.2.2+cu118
# Required for torch+cuda install
--extra-index-url https://download.pytorch.org/whl/cu118

# Core libraries
matplotlib
opencv-python
albumentations
scikit-learn
pillow>=10.0.0
tqdm
jupyter
ipykernel

# HuggingFace ecosystem
transformers
datasets

# Visualization
seaborn

# Deployment
gradio>=4.26.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.6

# For downloading model weights
gdown>=5.0.0